adds cordova-plugin-test framework to be able to run tests against the window object
