+++ package.json
@@ -41,6 +41,7 @@
         "test": "jasmine-node --color spec"
     },
     "devDependencies": {
-        "jasmine-node": "1.14.5"
+        "jasmine-node": "1.14.5",
+        "cordova-paramedic": "0.5.0"
     }
 }
